+++ src/router/routes.js
@@ -13,6 +13,8 @@ import BrowserExtension from '@/modules/wallets/pages/access/browser-extension/B
 import BrowserExtensionInstall from '@/modules/wallets/pages/access/browser-extension/install-extension/InstallExtension';
 import BrowserExtensionAccess from '@/modules/wallets/pages/access/browser-extension/access-extension/AccessExtension';
 import HowItWorks from '@/modules/how-it-works/HowItWorks';
+import PrivacyPolicy from '@/modules/privacy-policy/PrivacyPolicy';
+import TermsOfService from '@/modules/terms-of-service/TermsOfService';
 import WalletRoutes from '@/modules/wallets/routes.js';
 
 const routes = [
@@ -33,6 +35,22 @@ const routes = [
         path: 'how-it-works',
         name: 'HowItWorks',
         component: HowItWorks,
+        meta: {
+          requiresAuth: false
+        }
+      },
+      {
+        path: 'privacy-policy',
+        name: 'PrivacyPolicy',
+        component: PrivacyPolicy,
+        meta: {
+          requiresAuth: false
+        }
+      },
+      {
+        path: 'terms-of-service',
+        name: 'TermsOfService',
+        component: TermsOfService,
         meta: {
           requiresAuth: false
         }